requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-filter>=24.1
drf-spectacular>=0.27
python-dotenv>=1.0
whitenoise>=6.6
gunicorn>=21.2
psycopg2-binary>=2.9
